New teacher's getter

+++ src/main/java/ru/mail/polis/homework/io/blocking/StructureOutputStream.java
@@ -27,7 +27,7 @@ public class StructureOutputStream extends FileOutputStream {
         writeLong(structure.getId());
         writeString(structure.getName());
         writeSubStructures(structure.getSubStructures());
-        writeFloat((float) structure.getCoeff());
+        writeFloat(structure.getCoeff());
         writeFlags(new boolean[]{structure.isFlag1(), structure.isFlag2(), structure.isFlag3(), structure.isFlag4()});
         write(structure.getParam());
     }
